% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_extract_latent.R
\name{bake.step_extract_latent}
\alias{bake.step_extract_latent}
\title{Title}
\usage{
\method{bake}{step_extract_latent}(object, new_data, ...)
}
\arguments{
\item{object}{The updated step function that has been through prep()}

\item{new_data}{Tibble of data to be processed}

\item{...}{a}
}
\value{
a
}
\description{
Title
}
\examples{
a <- 1
}
