% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_if_gpu_available}
\alias{check_if_gpu_available}
\title{Check if GPU is available}
\usage{
check_if_gpu_available()
}
\value{
Command line message indicating whether GPU is available or not.
}
\description{
Check if GPU is available
}
\examples{
check_if_gpu_available()
}
