% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scarf_network.R
\name{classifier_network}
\alias{classifier_network}
\title{Classification Head torch module.}
\usage{
classifier_network(input_dim, n_classes, dropout = 0)
}
\arguments{
\item{input_dim}{Input dimension}

\item{n_classes}{Number of classes to classify}

\item{dropout}{Dropout layer parameter. Default: 0.0 (without dropout)}
}
\value{
A 'torch::nn_module' representing the classification head.
}
\description{
Classification Head torch module.
}
