% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dataset.R
\name{create_tensor_dataset}
\alias{create_tensor_dataset}
\title{Title}
\usage{
create_tensor_dataset(data)
}
\arguments{
\item{data}{Original data samples}
}
\value{
A 'torch::dataset' that contains the data.
}
\description{
Title
}
\examples{
X_fake <- matrix(runif(50*4), nrow=50, ncol=4)

my_dataset <- create_tensor_dataset(X_fake)

first_item <- my_dataset$.getitem(1)
print(first_item$x)
}
