% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dataset.R
\name{create_tensor_dataset_with_label}
\alias{create_tensor_dataset_with_label}
\title{Creation of a tensor dataset.}
\usage{
create_tensor_dataset_with_label(data, target)
}
\arguments{
\item{data}{Original data samples}

\item{target}{Original label samples}
}
\value{
A 'torch::dataset' that contains the data.
}
\description{
Creation of a tensor dataset.
}
