% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_representations.R
\name{downstream_prediction}
\alias{downstream_prediction}
\title{Title}
\usage{
downstream_prediction(
  df_test,
  target_column,
  exclude_columns,
  pretrained_model_path,
  classification_model_path,
  return_classification_report = FALSE
)
}
\arguments{
\item{df_test}{a}

\item{target_column}{a}

\item{exclude_columns}{a}

\item{pretrained_model_path}{a}

\item{classification_model_path}{a}

\item{return_classification_report}{a}
}
\value{
a
}
\description{
Title
}
\examples{
a <- 1
}
