% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_bundle}
\alias{load_bundle}
\title{Function to read and check a SCARF bundle}
\usage{
load_bundle(bundle_path, pretraining_type = "SCARF")
}
\arguments{
\item{bundle_path}{Path to file storing a "scarf_bundle" object}

\item{pretraining_type}{A \code{character} indicating the pretraining objective used by the model stored in the bundle. Default is \code{"SCARF"}. Available options are \code{["SCARF"]}.}
}
\value{
Pretrained model's weights and trained recipe for preprocessing
}
\description{
Function to read and check a SCARF bundle
}
