% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_scarf_bundle}
\alias{load_scarf_bundle}
\title{Function to read and check a SCARF bundle}
\usage{
load_scarf_bundle(bundle_path)
}
\arguments{
\item{bundle_path}{Path to file storing a "scarf_bundle" object}
}
\value{
Pretrained model's weights and trained recipe for preprocessing
}
\description{
Function to read and check a SCARF bundle
}
