% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_extract_latent.R
\name{prep.step_extract_latent}
\alias{prep.step_extract_latent}
\title{Title}
\usage{
\method{prep}{step_extract_latent}(x, training, info = NULL, ...)
}
\arguments{
\item{x}{The step_extract_latent object}

\item{training}{a}

\item{info}{a}

\item{...}{a}
}
\value{
a
}
\description{
Title
}
\examples{
a <- 1
}
