% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{prepare_scarf_data}
\alias{prepare_scarf_data}
\title{Prepare data for SCARF pretraining}
\usage{
prepare_scarf_data(
  dataframe_train,
  exclude_columns = NULL,
  create_validation = FALSE,
  validation_proportion = 0.1
)
}
\arguments{
\item{dataframe_train}{Train dataframe}

\item{exclude_columns}{Columns that the pretraining model should avoid (i.e target or ID columns)}

\item{create_validation}{Indicate whether a validation set should be created}

\item{validation_proportion}{Proportion of the training samples that will be used to create the validation set, if required.}
}
\value{
Preprocessed train dataset (and validation set if required) and the recipes::recipe used for preprocessing
}
\description{
Prepare data for SCARF pretraining
}
