% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{prepare_scarf_data_for_feature_extraction}
\alias{prepare_scarf_data_for_feature_extraction}
\title{Prepare data for SCARF feature extraction}
\usage{
prepare_scarf_data_for_feature_extraction(
  dataframe,
  trained_recipe,
  exclude_columns = NULL
)
}
\arguments{
\item{dataframe}{Dataframe for feature extraction}

\item{trained_recipe}{Trained recipes::recipe for preprocessing}

\item{exclude_columns}{Columns that the pretraining model should avoid (i.e target or ID columns)}
}
\value{
A torch::matrix representing the dataframe ready for feature extraction
}
\description{
Prepare data for SCARF feature extraction
}
