% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_parquet_data}
\alias{read_parquet_data}
\title{Read parquet data and prepare for training}
\usage{
read_parquet_data(filename_train, filename_test)
}
\arguments{
\item{filename_train}{Path to the training file}

\item{filename_test}{Path to the testing file}
}
\value{
An object type 'list' with training, validation and testing (data and labels) subsets
}
\description{
Read parquet data and prepare for training
}
