% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network_models.R
\name{scarf_encoder}
\alias{scarf_encoder}
\title{Basic encoder module.}
\usage{
scarf_encoder(in_dim, hidden_dim = 256, num_hidden = 4, dropout = 0)
}
\arguments{
\item{in_dim}{Number of input features.}

\item{hidden_dim}{Number of hidden or latent features.}

\item{num_hidden}{Number of blocks of layers of the encoder network.}

\item{dropout}{Dropout probability.}
}
\value{
A 'torch::nn_module' representing the encoder.
}
\description{
Basic encoder module.
}
