% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extractor.R
\name{scarf_feature_extractor}
\alias{scarf_feature_extractor}
\title{Title}
\usage{
scarf_feature_extractor(
  dataframe,
  pretrained_model_path,
  exclude_columns = NULL,
  batch_size = 32
)
}
\arguments{
\item{dataframe}{Dataframe to process and extract features}

\item{pretrained_model_path}{Path to the pretrained model used to extract representations}

\item{exclude_columns}{Columns that the pretraining model should avoid (i.e target or ID columns). Default = NULL}

\item{batch_size}{Batch size used during feature extraction. Default = 32}
}
\value{
TODO
}
\description{
Title
}
\examples{
a <- 1
}
