% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pretraining.R
\name{scarf_pretraining}
\alias{scarf_pretraining}
\title{Title}
\usage{
scarf_pretraining(path_train_data, path_test_data)
}
\arguments{
\item{path_train_data}{tdo}

\item{path_test_data}{tdo}
}
\value{
tdo
}
\description{
Title
}
\examples{
a <- 1
}
