% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_dataset.R
\name{create_tensor_dataset}
\alias{create_tensor_dataset}
\title{Title}
\usage{
create_tensor_dataset(data)
}
\arguments{
\item{data}{Original data samples}
}
\value{
A 'torch::dataset' that contains the data.
}
\description{
Title
}
