% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_representations.R
\name{downstream_prediction}
\alias{downstream_prediction}
\title{Title}
\usage{
downstream_prediction(
  df_test,
  pretrained_model_path,
  label_column,
  classification_model_path,
  exclude_columns = NULL,
  return_classification_report = FALSE
)
}
\arguments{
\item{df_test}{Test dataframe in which the classification model will predict each sample.}

\item{pretrained_model_path}{Path to the pretrained model.}

\item{classification_model_path}{Path to the classification model.}

\item{exclude_columns}{Columns that the classification model should ignore during training and inference (i.e target or ID columns). Default: NULL.}

\item{return_classification_report}{Indicate whether the user want to produce a classification report (TRUE) or just the predictions (FALSE). Default: FALSE.}

\item{target_column}{Column of the dataframe that store the sample's label.}
}
\value{
A "list" with the predicted label and probability score of each sample.
}
\description{
Title
}
\examples{
a <- 1
}
