% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pretraining.R
\name{scarf_fit}
\alias{scarf_fit}
\title{SCARF pretraining}
\usage{
scarf_fit(
  dataframe_train,
  exclude_columns = NULL,
  create_validation = FALSE,
  validation_proportion = 0.1,
  batch_size = 256,
  n_epochs = 1
)
}
\arguments{
\item{dataframe_train}{Train dataframe}

\item{exclude_columns}{Columns that the pretraining model should avoid (i.e target columns)}

\item{create_validation}{Indicate whether a validation set should be created}

\item{validation_proportion}{Proportion of the training samples that will be used to create the validation set, if required.}

\item{batch_size}{Batch size used during pretraining}

\item{n_epochs}{Number of pretraining epochs}
}
\value{
A pretrained RDS model
}
\description{
SCARF pretraining
}
\examples{
a <- 1
}
